Working loadgenerator with creates, updates and deletes.

+++ src/main/java/travelsampleloadgen/service/DocumentTemplate.java
@@ -8,11 +8,11 @@ import org.json.simple.parser.ParseException;
 public abstract class DocumentTemplate {
 long seed, revisionSeed;
 	
-	public void setSeed(long seed, int revison) {
+	public void setSeed(long seed, long revison) {
 		this.seed = seed;
 		Random rand = new Random(seed);
 		this.revisionSeed = rand.nextLong();
-		for(int i = 1;i < revison; i++) {
+		for(long i = 1;i < revison; i++) {
 			this.revisionSeed = rand.nextLong();
 		}
 	}
